remove updatetime configuration
Updatetime is used because CursorHold is by default used to determine
when to AutoSave. But different users prefer different
AutoSave behaviors. Some don't use CursorHold at all. Some
might prefer a different updatetime setting. Furthermore a too small
updatetime might break other plugins.

We should therefore not configure it for the user.
Instead a description on how to change it should be enough
to enable someone who wants to use a CursorHold event for autosave
to configure it himself.